First draft of the file reading/parsing component
It seems to be easier to parse the process description if we add some extra entries into the file. However, the instruction mentions that we need to read a process description file provided by the TA. I am not sure if we can also modify TA's file to fit our implementation. Did the prof mention this during class?